+++ apps/web/src/views/TradingReward/components/Leaderboard/index.tsx
@@ -39,7 +39,7 @@ const Leaderboard: React.FC<React.PropsWithChildren<LeaderboardProps>> = ({ camp
   }
 
   return (
-    <Box position="relative" style={{ zIndex: 1 }} mt="104px">
+    <Box id="leaderboard" position="relative" style={{ zIndex: 1 }} mt="104px">
       <Box>
         <Text textAlign="center" color="secondary" mb="16px" fontSize={['40px']} bold lineHeight="110%">
           {t('Leaderboard')}