+++ packages/pools/src/constants/pools/56.ts
@@ -14,6 +14,24 @@ export const livePools: SerializedPool[] = [
     isFinished: false,
   },
   {
+    sousId: 349,
+    stakingToken: bscTokens.cake,
+    earningToken: bscTokens.pendle,
+    contractAddress: '0x658136728Cf2D0845BEE552B07cA7FE82D1B7938',
+    poolCategory: PoolCategory.CORE,
+    tokenPerBlock: '0.01426',
+    version: 3,
+  },
+  {
+    sousId: 348,
+    stakingToken: bscTokens.cake,
+    earningToken: bscTokens.xalgo,
+    contractAddress: '0x4b5A6BA2cAf4B6FD04407513172D3e01c288f9bD',
+    poolCategory: PoolCategory.CORE,
+    tokenPerBlock: '0.07716',
+    version: 3,
+  },
+  {
     sousId: 347,
     stakingToken: bscTokens.cake,
     earningToken: bscTokens.tusd,
@@ -40,15 +58,6 @@ export const livePools: SerializedPool[] = [
     version: 3,
   },
   {
-    sousId: 343,
-    stakingToken: bscTokens.cake,
-    earningToken: bscTokens.edu,
-    contractAddress: '0x3d2d34Ea77B3702B7634C8D208feC5E08CEa88B6',
-    poolCategory: PoolCategory.CORE,
-    tokenPerBlock: '0.05425',
-    version: 3,
-  },
-  {
     sousId: 325,
     stakingToken: bscTokens.cake,
     earningToken: bscTokens.lvl,
@@ -66,6 +75,15 @@ export const livePools: SerializedPool[] = [
 
 // known finished pools
 const finishedPools = [
+  {
+    sousId: 343,
+    stakingToken: bscTokens.cake,
+    earningToken: bscTokens.edu,
+    contractAddress: '0x3d2d34Ea77B3702B7634C8D208feC5E08CEa88B6',
+    poolCategory: PoolCategory.CORE,
+    tokenPerBlock: '0.05425',
+    version: 3,
+  },
   {
     sousId: 342,
     stakingToken: bscTokens.cake,