+++ packages/tokens/src/1.ts
@@ -236,8 +236,8 @@ export const ethereumTokens = {
     ChainId.ETHEREUM,
     '0x56C03B8C4FA80Ba37F5A7b60CAAAEF749bB5b220',
     18,
-    'Canto',
     'CANTO',
+    'Canto',
     'https://tusd.io/',
   ),
 }