+++ packages/pools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pancakeswap/pools
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [e31475e6b]
+  - @pancakeswap/token-lists@0.0.8
+  - @pancakeswap/utils@2.0.1
+  - @pancakeswap/tokens@0.1.4
+
 ## 2.0.0
 
 ### Major Changes